@@ -28,18 +28,22 @@
   },
   "private": true,
   "dependencies": {
+    "@storybook/builder-webpack5": "6.5.13",
+    "@storybook/core-server": "6.5.13",
+    "@storybook/manager-webpack5": "6.5.13",
     "core-js": "3.11.0",
     "document-register-element": "1.14.10",
     "i18next": "^21.6.16",
     "inversify": "6.0.1",
     "lodash": "4.17.21",
-    "react": "17.0.2",
+    "react": "18.2.0",
     "react-data-table-component": "7.5.0",
-    "react-dom": "17.0.2",
+    "react-dom": "18.2.0",
     "react-i18next": "^11.16.7",
+    "react-is": "18.0.0",
     "react-router": "5.3.1",
     "react-router-config": "5.1.1",
-    "react-router-dom": "5.3.1",
+    "react-router-dom": "6.4.1",
     "react-router-named-routes": "0.0.23",
     "reflect-metadata": "0.1.13",
     "regenerator-runtime": "0.13.7",
@@ -50,52 +54,52 @@
     "@babel/preset-env": "7.13.15",
     "@babel/preset-react": "7.13.13",
     "@babel/preset-typescript": "7.13.0",
-    "@nrwl/cli": "12.10.1",
-    "@nrwl/cypress": "12.10.1",
-    "@nrwl/eslint-plugin-nx": "12.10.1",
-    "@nrwl/jest": "12.10.1",
-    "@nrwl/linter": "12.10.1",
-    "@nrwl/react": "12.10.1",
-    "@nrwl/storybook": "^12.10.1",
-    "@nrwl/tao": "12.10.1",
-    "@nrwl/web": "12.10.1",
-    "@nrwl/workspace": "12.10.1",
-    "@storybook/addon-essentials": "~6.3.0",
-    "@storybook/react": "~6.3.0",
+    "@nrwl/cli": "14.8.6",
+    "@nrwl/cypress": "14.8.6",
+    "@nrwl/eslint-plugin-nx": "14.8.6",
+    "@nrwl/jest": "14.8.6",
+    "@nrwl/linter": "14.8.6",
+    "@nrwl/react": "14.8.6",
+    "@nrwl/storybook": "14.8.6",
+    "@nrwl/web": "14.8.6",
+    "@nrwl/workspace": "14.8.6",
+    "@storybook/addon-essentials": "6.5.13",
+    "@storybook/react": "6.5.13",
     "@svgr/webpack": "^5.4.0",
-    "@testing-library/react": "11.2.6",
-    "@testing-library/react-hooks": "7.0.1",
-    "@types/jest": "27.0.2",
+    "@testing-library/react": "13.4.0",
+    "@types/jest": "28.1.8",
     "@types/lodash": "^4.14.168",
-    "@types/node": "12.20.10",
-    "@types/react": "17.0.3",
-    "@types/react-dom": "17.0.3",
+    "@types/node": "18.7.18",
+    "@types/react": "18.0.20",
+    "@types/react-dom": "18.0.6",
     "@types/react-router": "2.0.61",
     "@types/react-router-config": "5.0.6",
     "@types/react-router-dom": "5.3.3",
-    "@typescript-eslint/eslint-plugin": "4.31.2",
-    "@typescript-eslint/parser": "4.31.2",
-    "babel-jest": "27.2.3",
+    "@typescript-eslint/eslint-plugin": "5.41.0",
+    "@typescript-eslint/parser": "5.41.0",
+    "babel-jest": "28.1.3",
     "babel-loader": "8.1.0",
     "cypress": "^7.3.0",
     "dotenv": "10.0.0",
-    "eslint": "7.24.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-import": "2.26.0",
     "eslint-plugin-jest": "24.3.5",
-    "eslint-plugin-jsx-a11y": "6.4.1",
+    "eslint-plugin-jsx-a11y": "6.6.1",
     "eslint-plugin-prettier": "3.3.1",
-    "eslint-plugin-react": "7.23.1",
-    "eslint-plugin-react-hooks": "4.2.0",
+    "eslint-plugin-react": "7.31.8",
+    "eslint-plugin-react-hooks": "4.6.0",
     "husky": "6.0.0",
-    "jest": "27.2.3",
+    "jest": "28.1.3",
     "lerna": "4.0.0",
     "lint-staged": "10.5.4",
-    "prettier": "2.3.2",
-    "ts-jest": "27.0.5",
-    "ts-node": "9.1.1",
-    "typescript": "4.3.5",
+    "nx": "14.8.6",
+    "prettier": "2.7.1",
+    "react-test-renderer": "18.0.0",
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
+    "typescript": "4.8.4",
     "url-loader": "^3.0.0"
   },
   "husky": {
