@@ -1,2 +1,1 @@
-export * from './react-router';
-export * from './i18next';
+export * from '@dotdev/inversify';
